Change Container.Image type from Option[Int] to Option[Long]
This avoids integer overflow errors in JSON parsing

+++ client/src/main/scala/skuber/Container.scala
@@ -171,6 +171,6 @@ object Container {
 
   case class Image(
     names: List[String] = Nil,
-    sizeBytes: Option[Int] = None
+    sizeBytes: Option[Long] = None
   )
 }
